apprenticeship_applier/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.44.0
playwright-stealth==1.0.6
camoufox[geoip]==0.4.11
python-dotenv==1.0.1
2captcha-python==1.2.2
lxml==5.2.1
